fix: Make the wiggle consistent (#8194)
On busier pages or slower machines the wiggle is sometimes lost due to skipped frames and
a high chance of the skew calculation rounding to 0. This changes to using a counter
for the skew which provides a more consistent animation. Also sets the timeout
to 15ms instead of 10 to avoid double updates since 60fps is ~16 ms.